+++ models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	Id       int    `json:"-"`
+	Id       uint64 `json:"-"`
 	Name     string `json:"name"`
 	Username string `json:"username"`
 	Password string `json:"password"`
