+++ pkg/repository.go
@@ -3,15 +3,15 @@ package pkg
 import "Todo/models"
 
 type Repository interface {
-	CreateUser(user models.User) (int, error)
+	CreateUser(user models.User) (uint64, error)
 	GetUser(username, password string) (models.User, error)
 
-	CreateList(userId int, list models.TodoList) (int, error)
-	GetAllLists(userID int) ([]models.TodoList, error)
-	GetListByID(userID, listID int) (models.TodoList, error)
-	DeleteList(userID, listID int) error
-	UpdateList(userID, listID int, input models.UpdateListInput) error
+	CreateList(userId uint64, list models.TodoList) (uint64, error)
+	GetAllLists(userID uint64) ([]models.TodoList, error)
+	GetListByID(userID, listID uint64) (models.TodoList, error)
+	DeleteList(userID, listID uint64) error
+	UpdateList(userID, listID uint64, input models.UpdateListInput) error
 
-	CreateItems(listID int, item models.TodoItem) (int, error)
-	GetAllItems(userID, listID int) ([]models.TodoItem, error)
+	CreateItems(listID uint64, item models.TodoItem) (uint64, error)
+	GetAllItems(userID, listID uint64) ([]models.TodoItem, error)
 }
