+++ pkg/service.go
@@ -6,18 +6,18 @@ import (
 )
 
 type Service interface {
-	CreateUser(user models.User) (int, error)
+	CreateUser(user models.User) (uint64, error)
 	GenerateToken(username, password string) (string, error)
-	ParseToken(token string) (int, error)
+	ParseToken(token string) (uint64, error)
 	CreateCookieWithValue(value string) *http.Cookie
-	CheckUser(username, password string) (int, error)
+	CheckUser(username, password string) (uint64, error)
 
-	CreateList(userId int, list models.TodoList) (int, error)
-	GetAllLists(userID int) ([]models.TodoList, error)
-	GetListByID(userID, listID int) (models.TodoList, error)
-	DeleteList(userID, listID int) error
-	UpdateList(userID, listID int, input models.UpdateListInput) error
+	CreateList(userId uint64, list models.TodoList) (uint64, error)
+	GetAllLists(userID uint64) ([]models.TodoList, error)
+	GetListByID(userID, listID uint64) (models.TodoList, error)
+	DeleteList(userID, listID uint64) error
+	UpdateList(userID, listID uint64, input models.UpdateListInput) error
 
-	CreateItem(userID, listID int, item models.TodoItem) (int, error)
-	GetAllItems(userID, listID int) ([]models.TodoItem, error)
+	CreateItem(userID, listID uint64, item models.TodoItem) (uint64, error)
+	GetAllItems(userID, listID uint64) ([]models.TodoItem, error)
 }
